Do less arithmetic in merge_{hi,lo}
We achieve this by using a representation of the form (beginning index,
ending index) for the slices rather than (beginning index, length).
This mostly reduces the number of arithmetic operations in merge_hi,
while merge_lo remains more or less untouched.

This has a noticeable impact on the in the benchmark!